docs: auto generate by ci docs: Modifications of comments and full translations
Change Logs:

1）/egg/README.zh-CN.md：Line 66 to 68 (Translate English part to
Chinese in README.zh-CN.md）.

2）/egg/app/extend/request.js：Line 181 (Correct English syntaxes
error).

3）Comments added for:app/extend/context.js

4）Comments merged for:
4.1）app/extend/request.js
4.2）app/extend/response.js

Why to merge comments for properties?

It's that properties get/set is reguarded as a "whole property", this
means you ONLY need to write one snippet of comments on either "get" or "set"
(usually on the 1st part).

If you write two snippets of comments on both get and set, if you move
your mouse onto the property, you'll find either get or set will have a wrong intellisense tip of comments.